requirements.txt
interval==1.0.0
Pillow==7.1.0